Added hardware and HDL designs
The HDL contains two subdirectories, 1280x800 and 1280x1024, that correspond to
designs that output to those two VGA resolutions. The differences between the
files in the subdirectories are few and small, and if I figure out a way it
would be nicer to combine them into one project, presumably using some sort of
preprocessor directives.